backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
pyyaml==6.0.2
pynvml==11.5.3

# Auth (Phase 4)
python-jose[cryptography]==3.3.0
aiosmtplib==3.0.2

# Persistence (Phase 4)
aiosqlite==0.20.0

# RAG + memory (Phase 5) — Qdrant HTTP API is used directly via httpx,
# so we don't need qdrant-client. pypdf handles PDF extraction.
pypdf==5.1.0
python-multipart==0.0.17   # for file upload endpoints

# Per-user encrypted chat-history store (Phase 6):
# AES-256-GCM + HKDF-SHA256 over AUTH_SECRET_KEY (or HISTORY_SECRET_KEY).
cryptography==43.0.3
# Transitive dep of cryptography's CFFI bindings. Some sandbox/CI envs skip
# the transitive install; pinning explicitly turns the cryptic pyo3
# PanicException into a clean install-time resolution.
cffi>=1.17.1

# Cross-worker rate limiting + scheduler coordination (Phase 7).
# Optional at runtime: the backend falls back to in-memory if REDIS_URL is
# unset or connection fails.
redis[hiredis]==5.2.0
